Changed a lot this time.
First I put all the images into a Map so that all the images can be
accessed with a string key. These keys should be noted in the
DrawingSurface class. I also made the MovingImage take in a Map of
PImages so that I can access all the Player images directly from that
class.